+++ src/app/components/safety-action-confirm-dialog/safety-action-confirm-dialog.component.html
